Refactored the `obs.go` file so it now re-uses more code
I've extracted a couple of re-usable functions out of the functions that were
available in the original `obs.go` file. Now the functions of this file are
simple and straightforward. This re-usable functions have been placed in a new
file called `utils.go`. I've added tests for these new functions.