Add build revision to image version
OBS can rebuild an appliance when one of its packages is updated. This
forces the creation of a new appliance with the same version specified
inside of the kiwi file, but with a different build revision.

This commit makes obs2vagrant append the build revision to the
appliance version inside of the json file exposed to Vagrant. By doing
that the automatic rebuilds of OBS will be noticied by Vagrant.